requirements.txt
amqp==5.1.1
anyio==4.0.0
asgiref==3.7.2
astroid==2.15.6
async-timeout==4.0.3
billiard==4.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dill==0.3.7
Django==4.2.4
django-countries==7.5.1
django-environ==0.11.2
django-redis==5.3.0
djangorestframework==3.14.0
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.17.3
idna==3.4
isort==5.12.0
kombu==5.3.2
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psycopg2==2.9.7
pylint==2.17.5
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
